move: assets 폴더 이동 (#2)
