+++ src/app/StoreList.tsx
@@ -1,4 +1,4 @@
-import ChevronDownIcon from './assets/svg/ChevronDownIcon';
+import ChevronDownIcon from '../assets/svg/ChevronDownIcon';
 import { styles } from './styles/storeListStyles.css';
 
 export default function StoreList() {
