+++ src/assets/svg/LocationMarkerIcon.tsx
