+++ src/assets/svg/SearchIcon.tsx
